requirements.txt
asgiref==3.5.0
atomicwrites==1.4.0
attrs==21.4.0
colorama==0.4.4
Django==4.0.1
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
iniconfig==1.1.1
Markdown==3.3.6
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
PyJWT==2.3.0
pyparsing==3.0.7
pytest==7.0.1
pytest-django==4.5.2
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
tomli==2.0.1
tzdata==2021.5
